Add -keep_size option
It works if bsf > 1 and sets the size of outputted DjVu doc (not the
BG44 chunk in it) equal to the size of source image. This allows to
generate djvu documents with same size as input image which contains
only downscaled background layer without additional manipulations.
(Without it document geometry will be scaled down with bw44 layer).